basm: repeat directive has a start parameter

+++ cpclib-asm/src/error.rs
@@ -178,7 +178,7 @@ pub enum AssemblerError {
     RepeatIssue {
         error: Box<AssemblerError>,
         span: Z80Span,
-        repetition: usize
+        repetition: i32
     }
 }
